Refactor router to read from hooks state
