+++ package.json
@@ -26,7 +26,7 @@
         "d2": "31.8.2",
         "d2-api": "1.1.0",
         "d2-manifest": "1.0.0",
-        "d2-ui-components": "2.1.0-beta.1",
+        "d2-ui-components": "2.1.0-beta.3",
         "font-awesome": "4.7.0",
         "lodash": "4.17.20",
         "postcss-rtl": "1.7.3",
